Updated renderer to recent Python version + New model
Changed the import from importing everything from numpy to importing with alias since it messed up with the max function that was overwritten by numpy import. Somehow it didn't cause problems before, maybe because I used an old Python version, but now it is stable in 3.12.

Also it renders a star destroyer now.